ci: fix intermittent build failures and some other things (#80)
Some Android instrumentation tests were intermittently failing with a NullPointerException (NPE) when the system tried to access Resources.getConfiguration() on a mocked Context. This happened because the mock context returned a real Resources object, but the underlying Configuration was null in the test environment. This caused failures during test setup or when Android tried to bind the application context.

this was causing reporting to break (as it caused the app would crash and break the xml listener )

this pr fixes by refactoring the mocks to always return valid stuff. will add more docs in future


* feat: co pilot instructions

* feat: dorny actions summary

* fix: build time at start

* fix: fix

* ci: maybe fix coverage reports

* ci: maybe fix jacoco actions report error

* fix: maybe build

* ci: maybe fix concert tests

* test: maybe fix legacy tests in concert

* ci: maybe

* test: maybe other break fix

* fix: prep for windows build

for now just deletes problematic node_modules files

but could be used for whatever makes the build work on windows

* test: unfuck connectedAndroidTest by getting rid of the errant execution line

execution 'ANDROIDX_TEST_ORCHESTRATOR'

weird that it only broke cAT but not the instrument runner